python package folder restructure, python package creation
