+++ pyproject.toml
@@ -1,0 +1,59 @@
+[project]
+name = "inspired"
+description = "Inelastic Neutron Scattering Prediction for Instantaneous Results and Experimental Design"
+dynamic = ["version"]
+requires-python = ">=3.9"
+dependencies = [
+    "torch",
+    "phonopy",
+    "seekpath",
+    "matgl",
+    "chgnet"
+]
+license = { text = "GPL3.0" }
+
+[project.urls]
+homepage = "https://github.com/neutrons/inspired/"
+
+[build-system]
+requires = [
+  "setuptools",
+   "wheel",
+   "toml",
+   "versioningit"
+]
+build-backend = "setuptools.build_meta"
+
+[tool.versioningit.vcs]
+method = "git"
+default-tag = "0.2.0"
+
+[tool.versioningit.next-version]
+method = "minor"
+
+[tool.versioningit.format]
+distance = "{next_version}.dev{distance}"
+dirty = "{version}+d{build_date:%Y%m%d}"
+distance-dirty = "{next_version}.dev{distance}+d{build_date:%Y%m%d%H%M}"
+
+[tool.versioningit.write]
+file = "src/inspired/_version.py"
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.setuptools.package-data]
+"*" = ["*.yml","*.yaml","*.ini"]
+
+[project.gui-scripts]
+inspired = "inspired.inspired:gui"
+
+[tool.pylint]
+max-line-length = 120
+disable = ["too-many-locals",
+           "too-many-statements",
+           "too-many-instance-attributes",
+           "too-many-arguments",
+           "duplicate-code"
+]
+extension-pkg-whitelist = ["PyQt5"]
