+++ src/inspired/latent_space_model.py
