+++ src/inspired/pandas_model.py
