+++ src/inspired/set_paths.ui
